git subrepo clone --force git@github.com:Cofyc/perl6-redis.git github/Cofyc/perl6-redis
subrepo:
  subdir:   "github/Cofyc/perl6-redis"
  merged:   "0c5cdbe149"
upstream:
  origin:   "git@github.com:Cofyc/perl6-redis.git"
  branch:   "master"
  commit:   "0c5cdbe149"
git-subrepo:
  version:  "0.3.1"
  origin:   "git@github.com:ingydotnet/git-subrepo.git"
  commit:   "a7ee886"

+++ github/Cofyc/perl6-redis/t/02-pub&sub.t
@@ -4,11 +4,23 @@ use lib <t lib>;
 use Redis;
 use Test;
 
-my $r = Redis.new("127.0.0.1:63790", decode_response => True);
-$r.auth('20bdfc8e73365b2fde82d7b17c3e429a9a94c5c9');
-$r.flushall;
-
 plan 1;
 
-# TODO 
-is-deeply $r.publish("queue", "data"), 0;
+use Test::SpawnRedisServer;
+
+if SpawnRedis() -> $proc {
+    LEAVE {
+        $proc.kill('INT');
+    }
+
+    my $r = Redis.new("127.0.0.1:63790", decode_response => True);
+    $r.auth('20bdfc8e73365b2fde82d7b17c3e429a9a94c5c9');
+    $r.flushall;
+
+
+    # TODO
+    is-deeply $r.publish("queue", "data"), 0;
+}
+else {
+    skip-rest "no redis-server";
+}
