+++ github/Cofyc/perl6-redis/t/02-scripting.t
@@ -4,24 +4,37 @@ use lib <t lib>;
 use Redis;
 use Test;
 
-my $r = Redis.new("127.0.0.1:63790", decode_response => True);
-$r.auth('20bdfc8e73365b2fde82d7b17c3e429a9a94c5c9');
-$r.flushall;
+use Test::SpawnRedisServer;
 
-if $r.info<redis_version> gt "2.6" {
-    plan 5;
+if SpawnRedis() -> $proc {
+    LEAVE {
+        $proc.kill('INT');
+    }
 
-    $r.set("key", 1);
-    is-deeply $r.eval("return redis.call('get', 'key')", 0), "1";
+    my $r = Redis.new("127.0.0.1:63790", decode_response => True);
+    $r.auth('20bdfc8e73365b2fde82d7b17c3e429a9a94c5c9');
+    $r.flushall;
 
-    is-deeply $r.script_flush(), True;
+    if $r.info<redis_version> gt "2.6" {
+        plan 5;
 
-    dies-ok { $r.script_kill() };
+        $r.set("key", 1);
+        is-deeply $r.eval("return redis.call('get', 'key')", 0), "1";
 
-    my $sha = $r.script_load("return redis.call('get', 'key')");
-    is-deeply $r.script_exists($sha, "unknown"), [1, 0];
+        is-deeply $r.script_flush(), True;
 
-    is-deeply $r.evalsha($sha, 0), "1";
-} else {
-    done;
+        dies-ok { $r.script_kill() };
+
+        my $sha = $r.script_load("return redis.call('get', 'key')");
+        is-deeply $r.script_exists($sha, "unknown"), [1, 0];
+
+        is-deeply $r.evalsha($sha, 0), "1";
+    } else {
+        plan 1;
+        skip-rest "scripting not available on this redis-server";
+    }
+}
+else {
+    plan 1;
+    skip-rest "no redis-server";
 }
