+++ github/Cofyc/perl6-redis/t/02-server.t
@@ -4,13 +4,25 @@ use lib <t lib>;
 use Redis;
 use Test;
 
-my $r = Redis.new("127.0.0.1:63790", decode_response => True);
-$r.auth('20bdfc8e73365b2fde82d7b17c3e429a9a94c5c9');
+use Test::SpawnRedisServer;
 
 plan 3;
 
-is-deeply $r.flushall(), True;
+if SpawnRedis() -> $proc {
+    LEAVE {
+        $proc.kill('INT');
+    }
 
-is-deeply $r.flushdb(), True;
+    my $r = Redis.new("127.0.0.1:63790", decode_response => True);
+    $r.auth('20bdfc8e73365b2fde82d7b17c3e429a9a94c5c9');
 
-ok $r.info.WHAT === Hash;
+
+    is-deeply $r.flushall(), True;
+
+    is-deeply $r.flushdb(), True;
+
+    ok $r.info.WHAT === Hash;
+}
+else {
+    skip-rest "no redis-server";
+}
